% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_death_data.R
\name{crosswalk_death_data}
\alias{crosswalk_death_data}
\title{Crosswalk Death Data File Years}
\usage{
crosswalk_death_data(df)
}
\value{
A data frame
}
\description{
Death Certificate Statistical files changed considerably after the Washington State Department of Health changed collection systems
from the Electronic Death Registration System (EDRS) to the Washington Health and Life Event Systems (WHALES). Specifically, the death certificate data
is considerably different from 2010-2015 versus 2016-onwards and requires considerable reformatting to preserve longitudinal analysis across these time periods.
}
