% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_death_data.R
\name{load_death_data}
\alias{load_death_data}
\title{Load Death Data}
\usage{
load_death_data(folder, include_prelim, file_years = NULL)
}
\arguments{
\item{folder}{(String) The folder path where your organzation stores downloaded death certificate statistical files.}

\item{include_prelim}{(T|F) Option to laod preliminary death certificate statistical files in addition to finalized files.}
}
\value{
A list of data frames (labelled by file year) containing death certificate statistical files
}
\description{
Load Death Certificate Files downloaded from Secure Access Washington (SAW) WA DOH CHS Data Files.
}
