% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walhj.deaths-package.R
\docType{package}
\name{walhj.deaths-package}
\alias{walhj.deaths}
\alias{walhj.deaths-package}
\title{walhj.deaths: Load, Clean, and Analyze Death Certificate data}
\description{
This R packages provides custom R functions to support the loading, cleaning, and analysis of Death Certificate data by Washington Local Health Jurisdictions (LHJs).
}
\author{
\strong{Maintainer}: Tyler Bonnell \email{Tyler.Bonnell@co.snohomish.wa.us}

Authors:
\itemize{
  \item Jacob Armitage \email{Jacob.Armitage@co.thurston.wa.us}
  \item Trey Evans \email{Trey.Evans@kitsappublichealth.org}
}

}
\keyword{internal}
